requirements.txt
anthropic>=0.40.0,<1.0
feedparser>=6.0.10,<7.0
PyYAML>=6.0,<7.0
